+++ src/lem/coroutine/mod.rs
@@ -1,2 +1,3 @@
 mod eval;
+mod synthesis;
 pub mod toplevel;
